refactor: Move components under posts
