--- src/types/tag.ts
@@ -1,5 +1,0 @@
-export type TagProps = {
-  appType: string;
-  category: string;
-  status: string;
-};
